+++ octoprint_flashforge/__init__.py
@@ -16,9 +16,12 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 
 	VENDOR_IDS = {0x0315: "PowerSpec", 0x2a89: "Dremel", 0x2b71: "FlashForge"}
 	PRINTER_IDS = {
-		"PowerSpec": {0x0001: "Ultra 3DPrinter"},
-		"Dremel": {0x8889: "Dremel IdeaBuilder"},
-		"FlashForge": {0x0001: "Dreamer", 0x000A: "Dreamer NX", 0x0002: "Finder v1", 0x0005: "Inventor", 0x0007: "Finder v2", 0x00ff: "PowerSpec Ultra"}}
+		"PowerSpec": {0x0001: "Ultra 3DPrinter (C)"},
+		"Dremel": {0x8889: "Dremel IdeaBuilder 3D20"},
+		"FlashForge": {0x0001: "Dreamer", 0x000A: "Dreamer NX", 0x0002: "Finder v1", 0x0005: "Inventor", 0x0007: "Finder v2",
+					   0x0009: "Guider IIs",
+					   0x00e7: "Creator Max", 0x00ee: "Finder v2.12",
+					   0x00f6: "PowerSpec Ultra 3DPrinter (B)", 0x00ff: "PowerSpec Ultra 3DPrinter (A)"}}
 	FILE_PACKET_SIZE = 1024
 
 
@@ -32,6 +35,7 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 		self._currentFile = None
 		self._upload_percent = 0
 		self._vendor_id = 0
+		self._vendor_name = ""
 		self._device_id = 0
 		# FlashForge friendly default connection settings
 		self._conn_settings = {
@@ -57,6 +61,8 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 	def get_settings_defaults(self):
 		# put your plugin's default settings here
 		return dict(
+			ledStatus=1,
+			ledColor=[255, 255, 255]
 		)
 
 
@@ -65,6 +71,8 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 		# Define your plugin's asset files to automatically include in the
 		# core UI here.
 		return dict(
+			js=["js/flashforge.js", "js/color-picker.min.js"],
+			css=["css/color-picker.min.css"]
 		)
 
 
@@ -102,20 +110,39 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 				except:
 					device_name = 'unknown'
 				self._logger.debug("Found device '{}' with Vendor ID: {:#06X}, USB ID: {:#06X}".format(device_name, vendor_id, device_id))
+				# get USB interface details to diagnose connectivity issues
+				for configuration in device.iterConfigurations():
+					for interface in configuration:
+						for setting in interface:
+							self._logger.debug(
+								" setting number: 0x{:02x}, class: 0x{:02x}, subclass: 0x{:02x}, protocol: 0x{:02x}, #endpoints: {}, descriptor: {}".format(
+								setting.getNumber(), setting.getClass(), setting.getSubClass(),
+								setting.getProtocol(), setting.getNumEndpoints(), setting.getDescriptor()))
+							for endpoint in setting:
+								self._logger.debug(
+									"  endpoint address: 0x{:02x}, attributes: 0x{:02x}, max packet size: {}".format(
+									endpoint.getAddress(), endpoint.getAttributes(),
+									endpoint.getMaxPacketSize()))
+
 				if vendor_id in self.VENDOR_IDS:
 					vendor_name = self.VENDOR_IDS[vendor_id]
 					if device_id in self.PRINTER_IDS[vendor_name]:
 						self._logger.info("Found a {} {}".format(vendor_name, self.PRINTER_IDS[vendor_name][device_id]))
 						self._vendor_id = vendor_id
+						self._vendor_name = vendor_name
 						self._device_id = device_id
 						break
 					else:
 						raise flashforge.FlashForgeError("Found an unsupported {} printer '{}' with USB ID: {:#06X}".format(vendor_name, device_name, device_id))
+
 		return self._device_id != 0
 
 
-	# Main serial connection hook - create our printer connection
 	def printer_factory(self, comm, port, baudrate, read_timeout, *args, **kwargs):
+		""" OctoPrint hook - Called when creating printer connection
+
+			Test for presence of a supported printer and then try to connect
+		"""
 		if not port == "AUTO":
 			return None
 
@@ -127,11 +154,16 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 		return serial_obj
 
 
-	# Supported file extensions for SD upload
 	def get_extension_tree(self, *args, **kwargs):
+		""" OctoPrint hook - Return supported file extensions for SD upload
+
+			Note not called when printer connects, only when starting up and when the printer disconnects
+		"""
+		self._logger.debug("get_extension_tree()")
 		return dict(
 			machinecode=dict(
-				gx=["gx"]
+				g3drem=["g3drem"],	# Dremel
+				gx=["gx"]			# Every other FlashForge based printer
 			)
 		)
 
@@ -152,6 +184,11 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 		self._serial_obj = None
 
 
+	def valid_command(self, command):
+		gcode = command.split(b' ', 1)[0]
+		return (gcode[0] in b"GM") and gcode not in [b"M117"]
+
+
 	# Called when gcode commands are being placed in the queue by OctoPrint:
 	# Mostly important for control panel or translating and printing non FlashPrint file directly from OctoPrint
 	def rewrite_gcode(self, comm_instance, phase, cmd, cmd_type, gcode, *args, **kwargs):
@@ -164,6 +201,8 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 
 			self._logger.debug("rewrite_gcode(): gcode:{}, cmd:{}".format(gcode, cmd))
 
+			#TODO: filter M146 and other commands? when printing from SD because they cause comms to hang
+
 			# M20 list SD card, M21 init SD card - do not do if we are busy, seems to cause issues
 			if (gcode == "M20" or gcode == "M21") and not self._serial_obj.is_ready():
 				cmd = []
@@ -174,7 +213,7 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 				if comm_instance.isCancelling():
 					cmd = []
 
-			# M26 is sent by OctoPrint during Sd prints:
+			# M26 is sent by OctoPrint during SD prints:
 			# M26 in Marlin = set SD card position : FlashForge = cancel
 			elif gcode == "M26":
 				# M26 S0 generated during OctoPrint cancel - use it to send cancel
@@ -222,7 +261,7 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 			elif gcode == "M110":
 				cmd = []
 
-			# also get printer status when connecting
+			# also get SD status when connecting
 			elif gcode == "M115":
 				cmd = [("M27", "sd_status_polling"), (cmd, cmd_type)]
 
@@ -230,11 +269,14 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 			elif gcode == "M119":
 				cmd = []
 
+			elif gcode == "M146":
+				cmd = []
+
 			# M400 is sent by OctoPrint on cancel:
 			# M400 in Marlin = wait for moves to finish : Flashforge = ? - instead send something inert so on_M400_sent
 			# is triggered in OctoPrint
 			elif gcode == "M400":
-				cmd = "M27"
+				cmd = [("M27", "sd_status_polling")]
 
 			if cmd == []:
 				self._logger.debug("rewrite_gcode(): dropping command")
@@ -260,11 +302,11 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 			error = "could not start tx"
 
 			# make sure heaters are off
-			self._serial_obj.sendcommand("M104 S0 T0")
-			self._serial_obj.sendcommand("M104 S0 T1")
-			self._serial_obj.sendcommand("M140 S0")
+			self._serial_obj.sendcommand(b"M104 S0 T0")
+			self._serial_obj.sendcommand(b"M104 S0 T1")
+			self._serial_obj.sendcommand(b"M140 S0")
 
-			ok, answer = self._serial_obj.sendcommand("M28 {} 0:/user/{}".format(file_size, remote_name), 5000)
+			ok, answer = self._serial_obj.sendcommand(b"M28 %d 0:/user/%s" % (file_size, remote_name.encode()), 5000)
 			if not ok:
 				error = "file transfer not started {}".format(answer)
 			else:
@@ -274,12 +316,12 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 			try:
 				while chunk_start_index < file_size:
 					chunk_end_index = min(chunk_start_index + self.FILE_PACKET_SIZE, file_size)
-					chunk = gcode[chunk_start_index:chunk_end_index]
+					chunk = bgcode[chunk_start_index:chunk_end_index]
 					if not chunk:
 						error = "unexpected eof"
 						break
 
-					if self._serial_obj.writeraw(chunk):
+					if self._serial_obj.writeraw(chunk, False):
 						upload_percent = 100.0 * chunk_end_index / file_size
 						self.upload_percent = int(upload_percent)
 						self._logger.debug("Sent: %.2f%% %d/%d" % (self.upload_percent, chunk_end_index, file_size))
@@ -290,10 +332,10 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 					chunk_start_index += self.FILE_PACKET_SIZE
 
 				if not error:
-					result, response = self._serial_obj.sendcommand("M29", 10000)
-					if result and "CMD M28" in response:
+					result, response = self._serial_obj.sendcommand(b"M29", 10000)
+					if result and b"CMD M28" in response:
 						response = self._serial_obj.readraw(1000)
-					if result and "failed" not in response:
+					if result and b"failed" not in response:
 						sd_upload_succeeded(filename, remote_name, 10)
 					else:
 						error = "File transfer incomplete"
@@ -309,22 +351,22 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 				raise flashforge.FlashForgeError(error)
 
 			self._serial_obj.makeexclusive(False)
-			# NB M23 select will also trigger a print on Flashforge
-			self._comm.selectFile("0:/user/{}\r\n".format(remote_name), True)
+			# NB M23 select will also trigger a print on FlashForge
+			self._comm.selectFile("0:/user/%s\r\n" % remote_name, True)
 			# TODO: need to set the correct file size for the progress indicator
 
 
 		import threading
 
-		gcode = ""
+		bgcode = b""
 		file_size = 0
 		# unfortunately we cannot get the list of files on the SD card from FlashForge so we just name the remote file
 		# the same as the source and hope for the best
 		remote_name = filename
 
-		file = open(path, "r")
-		gcode = file.read()
-		file_size = len(gcode)
+		file = open(path, "rb")
+		bgcode = file.read()
+		file_size = len(bgcode)
 		file.close()
 
 		self._logger.info("Starting SDCard upload from {} to {}".format(filename, remote_name))
@@ -342,6 +384,8 @@ class FlashForgePlugin(octoprint.plugin.SettingsPlugin,
 # ("OctoPrint-PluginSkeleton"), you may define that here. Same goes for the other metadata derived from setup.py that
 # can be overwritten via __plugin_xyz__ control properties. See the documentation for that.
 __plugin_name__ = "FlashForge Plugin"
+__plugin_pythoncompat__ = ">=2.7,<4"
+
 
 def __plugin_load__():
 	global __plugin_implementation__